null check

+++ portal/portal-service-impl/impl/src/java/org/sakaiproject/portal/service/PortalServiceImpl.java
@@ -1123,7 +1123,7 @@ public class PortalServiceImpl implements PortalService, Observer
 	@Override
 	public void removeSitesfromPinnedAndRecent(String userId, List<String> siteIds) {
 
-		if (StringUtils.isBlank(userId) || siteIds.isEmpty()) return;
+		if (StringUtils.isBlank(userId) || siteIds == null || siteIds.isEmpty()) return;
 
 		withPortalNavLock(userId, () -> {
 			pinnedSiteRepository.deleteByUserIdAndSiteIds(userId, siteIds);
